Do not force delete non-trashed items

+++ app/Http/Controllers/Api/Admin/UsersController.php
@@ -136,13 +136,11 @@ class UsersController extends Controller
             abort(400);
         }
 
-        if ($request->input('destroy') === 'true') {
-            $user = User::withTrashed()->findOrFail($id);
+        $user = User::withTrashed()->findOrFail($id);
 
+        if ($request->input('destroy') === 'true' && $user->trashed()) {
             $user->forceDelete();
         } else {
-            $user = User::findOrFail($id);
-
             $user->delete();
         }
     }
